Changed the way for computing future size. Modified in README.md. No changes in strategy and structure. Based on initial version on Sep. 14th of 2023.
Fixed some bugs in file referring and description parts. No changes in strategy and structure. Based on initial version on Sep. 14th of 2023.

In this version, a fixed alpha=5 is adopted for LASSO regression, with simple AR(1) and GARCH(1, 1) models for the equilibrium level and volatility. No stationary test will be conducted for the residual series constructed by LASSO regression. Also, no cross-validation will be conducted for the signal lower boundary. The boundary is specified simply by mean +/- 0.5 * sd.